requirements.txt
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
flask-swagger-ui==4.11.1
google-api-core==2.11.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-secret-manager==2.10.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.4
